completes force_login implementation; cleans up login code in home page
