+++ pages/4_Analytics.py
@@ -9,6 +9,9 @@ from datetime import datetime
 import pandas as pd
 import function
 import log_utils as log
+from auth_utils import force_login
+
+force_login()
 
 function.wide_space_default()
